requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scipy>=1.10.0
timm>=0.9.0
Pillow>=9.5.0
tqdm>=4.65.0
scikit-learn>=1.3.0
tensorboard>=2.13.0
matplotlib>=3.7.0
pyyaml>=6.0
ninja>=1.11.0
pytest>=7.3.0
black>=23.0.0
flake8>=6.0.0
# Enhanced MADR-GAN dependencies
faiss-cpu>=1.7.0
open-clip-torch>=2.20.0